payment_app/requirements.txt
confluent-kafka==1.8.2
psycopg2-binary==2.9.3